feat(pixiv-utils): 1.6.46, add options for slightly faster tags block/highlight
enabled by default.

it'll just make them return immediately with the first matching tag for highlighting/blocking.

the cons is being unable to see the full list of tags that would've triggered them via the tooltip, but assuming you don't care to know them, this way is indeed faster compute-wise.